prefactor: extract to GGardenCtlCommand
