+++ frontend/src/components/ShootDetails/GShootAccessCard.vue
@@ -89,7 +89,7 @@ SPDX-License-Identifier: Apache-2.0
       inset
     />
 
-    <g-gardenctl-commands
+    <g-shoot-gardenctl-commands
       v-if="isGardenctlTileVisible"
     />
   </g-list>
@@ -121,7 +121,7 @@ import {
 } from '@/composables/useShootItem'
 import { useProvideSeedHelper } from '@/composables/useSeedHelper'
 
-import GGardenctlCommands from './GGardenctlCommands.vue'
+import GShootGardenctlCommands from './GShootGardenctlCommands.vue'
 import GShootKubeconfig from './GShootKubeconfig.vue'
 import GShootAdminKubeconfig from './GShootAdminKubeconfig.vue'
 import GTerminalShortcutsTile from './GTerminalShortcutsTile.vue'
